Moved autogenerated files from queries to Scaffold
